+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @glissade/core
 
+## 0.56.0-pre.0
+
 ## 0.55.0
 
 ### Minor Changes